feat(api): enhance credit service with early bird user check
- Added a logger to the CreditService for improved logging.
- Introduced a new method to determine if a user is an early bird based on their subscription status.
- Updated syncTokenCreditUsage method to skip credit billing for early bird users when applicable.
- Refactored skill-invoker service to streamline credit billing checks and usage reporting.
